refactor: separating the token for helper and importing
